Update models.py (#93)

+++ project/ai_operations/models.py
@@ -25,7 +25,7 @@ class Node(models.Model):
 class Component(models.Model):
     displayed_name = models.CharField(max_length=255, default="")
     description = models.CharField(max_length=255, default="")
-    idx = models.BigIntegerField(default=0)
+    idx = models.IntegerField(default=0)
     category = models.CharField(max_length=255, default="")
     node_name = models.CharField(max_length=255)
     node_type = models.CharField(max_length=255, default="general")
